pyproject.toml
[project]
name = "langraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "ipykernel>=6.29.5",
    "langchain-google-genai>=2.1.4",
    "langchain[anthropic]>=0.3.25",
    "langgraph>=0.4.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
